rename setPicture to uploadPicture
