+++ server/api/user/index.js
@@ -9,7 +9,7 @@ import revokeRole from './revokeRole';
 import listUsers from './list';
 import deactivateUser from './deactivate';
 import activateUser from './activate';
-import setUserPicture from './setPicture';
+import uploadUserPicture from './uploadPicture';
 import deleteUserPicture from './deletePicture';
 import forgotPassword from './forgotPassword';
 import resetPassword from './resetPassword';
@@ -28,7 +28,7 @@ router.post('user.revokeRole', '/user.revokeRole', revokeRole);
 router.post('user.list', '/user.list', listUsers);
 router.post('user.deactivate', '/user.deactivate', deactivateUser);
 router.post('user.activate', '/user.activate', activateUser);
-router.post('user.setPicture', '/user.setPicture', setUserPicture);
+router.post('user.uploadPicture', '/user.uploadPicture', uploadUserPicture);
 router.post('user.deletePicture', '/user.deletePicture', deleteUserPicture);
 router.post('user.forgotPassword', '/user.forgotPassword', forgotPassword);
 router.post('user.resetPassword', '/user.resetPassword', resetPassword);
