+++ server/api/user/uploadPicture/__tests__/coyote.jpg
