+++ server/api/user/uploadPicture/__tests__/runner.png
