+++ server/api/user/uploadPicture/controller.js
