+++ server/api/user/uploadPicture/controller.test.js
@@ -11,7 +11,7 @@ import deletePermissionAssignment from '../revokePermission/service';
 import deleteUser from '../delete/service';
 import deleteUserPicture from '../deletePicture/service';
 
-describe('api/user.setPicture', () => {
+describe('api/user.uploadPicture', () => {
   let ctx;
 
   beforeAll(async () => {
@@ -26,7 +26,7 @@ describe('api/user.setPicture', () => {
   describe('unauthorized user', () => {
     test('returns error pretending resource does not exist', async () => {
       const res = await request(server)
-        .post('/api/user.setPicture')
+        .post('/api/user.uploadPicture')
         .send();
       expect(res.status).toBe(200);
       expect(res.body).toMatchObject({
@@ -106,7 +106,7 @@ describe('api/user.setPicture', () => {
 
     test('sets user profile picture', async () => {
       let res = await request(server)
-        .post('/api/user.setPicture')
+        .post('/api/user.uploadPicture')
         .set('Authorization', `Bearer ${wileSession.accessToken}`)
         .attach('picture', path.resolve(__dirname, './__tests__/runner.png'))
         .field('id', runner.id);
@@ -129,7 +129,7 @@ describe('api/user.setPicture', () => {
 
     test('returns error when `id` contains invalid characters', async () => {
       const res = await request(server)
-        .post('/api/user.setPicture')
+        .post('/api/user.uploadPicture')
         .set('Authorization', `Bearer ${wileSession.accessToken}`)
         .attach('picture', path.resolve(__dirname, './__tests__/coyote.jpg'))
         .field('id', '507f1f77bcf86cd79943901@');
@@ -148,7 +148,7 @@ describe('api/user.setPicture', () => {
 
     test('returns error when `id` contains more than 24 characters', async () => {
       const res = await request(server)
-        .post('/api/user.setPicture')
+        .post('/api/user.uploadPicture')
         .set('Authorization', `Bearer ${wileSession.accessToken}`)
         .attach('picture', path.resolve(__dirname, './__tests__/coyote.jpg'))
         .field('id', '507f1f77bcf86cd7994390112');
@@ -167,7 +167,7 @@ describe('api/user.setPicture', () => {
 
     test('returns error when `id` contains less than 24 characters', async () => {
       const res = await request(server)
-        .post('/api/user.setPicture')
+        .post('/api/user.uploadPicture')
         .set('Authorization', `Bearer ${wileSession.accessToken}`)
         .attach('picture', path.resolve(__dirname, './__tests__/coyote.jpg'))
         .field('id', '507f1f77bcf86cd79943901');
@@ -186,7 +186,7 @@ describe('api/user.setPicture', () => {
 
     test('returns error when `id` is unspecified', async () => {
       const res = await request(server)
-        .post('/api/user.setPicture')
+        .post('/api/user.uploadPicture')
         .set('Authorization', `Bearer ${wileSession.accessToken}`)
         .attach('picture', path.resolve(__dirname, './__tests__/coyote.jpg'));
       expect(res.status).toBe(200);
@@ -204,7 +204,7 @@ describe('api/user.setPicture', () => {
 
     test('returns error when payload contains unknown properties', async () => {
       const res = await request(server)
-        .post('/api/user.setPicture')
+        .post('/api/user.uploadPicture')
         .set('Authorization', `Bearer ${wileSession.accessToken}`)
         .attach('picture', path.resolve(__dirname, './__tests__/coyote.jpg'))
         .field('id', '507f1f77bcf86cd799439011')
@@ -224,7 +224,7 @@ describe('api/user.setPicture', () => {
 
     test('returns error with invalid permission scope', async () => {
       const res = await request(server)
-        .post('/api/user.setPicture')
+        .post('/api/user.uploadPicture')
         .set('Authorization', `Bearer ${runnerSession.accessToken}`)
         .attach('picture', path.resolve(__dirname, './__tests__/coyote.jpg'))
         .field('id', wile.id);
@@ -240,7 +240,7 @@ describe('api/user.setPicture', () => {
 
     test('can set their own profile picture', async () => {
       let res = await request(server)
-        .post('/api/user.setPicture')
+        .post('/api/user.uploadPicture')
         .set('Authorization', `Bearer ${runnerSession.accessToken}`)
         .attach('picture', path.resolve(__dirname, './__tests__/runner.png'))
         .field('id', runner.id);
