+++ server/api/user/uploadPicture/index.js
